+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## [2025.2.5.1]
 
 <cite>Release contributors</cite>
-- 3 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.2.5.1+author%3Amlytvyn+is%3Apr)
+- 4 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.2.5.1+author%3Amlytvyn+is%3Apr)
+
+### `Project Import` enhancements
+- Automatically download  3rd-part library dependencies [#1669](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1669)
 
 ### `hAC` enhancements
 - Added collapsible groups for connection dialog to target low-resolution screens [#1664](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1664)